+++ src/main/java/nc/noumea/mairie/ptg/service/impl/ExportPaiePrimeService.java
@@ -26,28 +26,28 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 
 	@Autowired
 	private IExportPaieRepository exportPaieRepository;
-	
+
 	@Autowired
 	private HelperService helperService;
-	
+
 	@Override
 	public List<Sppprm> exportPrimesJourToPaie(List<Pointage> pointagesOrderedByDateAsc) {
-		
+
 		List<Sppprm> modifiedOrAddedSppprm = new ArrayList<Sppprm>();
-		
+
 		for (Pointage ptg : pointagesOrderedByDateAsc) {
-			
+
 			if (ptg.getTypePointageEnum() != RefTypePointageEnum.PRIME
-				|| ptg.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPPRM)
+					|| ptg.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPPRM)
 				continue;
-			
+
 			// Fetch or create Sppprm
 			Sppprm prm = findOrCreateSppprmRecord(modifiedOrAddedSppprm, ptg.getIdAgent(), ptg.getDateDebut(), ptg
 					.getRefPrime().getNoRubr(), ptg.getQuantite());
 
 			if (prm == null)
 				continue;
-			
+
 			switch (ptg.getRefPrime().getTypeSaisie()) {
 				case NB_HEURES:
 				case PERIODE_HEURES:
@@ -60,7 +60,7 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 					break;
 
 			}
-			
+
 			// If quantity is 0, remove this record from the list of Sppprm
 			if (prm.getNbPrime() == 0) {
 				modifiedOrAddedSppprm.remove(prm);
@@ -68,28 +68,28 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 			}
 
 		}
-		
+
 		return modifiedOrAddedSppprm;
 	}
-	
+
 	@Override
 	public List<Sppprm> exportPrimesCalculeesJourToPaie(List<PointageCalcule> pointagesCalculesOrderedByDateAsc) {
 
 		List<Sppprm> modifiedOrAddedSppprm = new ArrayList<Sppprm>();
-		
+
 		for (PointageCalcule ptgC : pointagesCalculesOrderedByDateAsc) {
-			
+
 			if (ptgC.getTypePointageEnum() != RefTypePointageEnum.PRIME
-				|| ptgC.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPPRM)
+					|| ptgC.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPPRM)
 				continue;
-			
+
 			// Fetch or create Sppprm
 			Sppprm prm = findOrCreateSppprmRecord(modifiedOrAddedSppprm, ptgC.getIdAgent(), ptgC.getDateDebut(), ptgC
 					.getRefPrime().getNoRubr(), ptgC.getQuantite());
-		
+
 			if (prm == null)
 				continue;
-			
+
 			switch (ptgC.getRefPrime().getTypeSaisie()) {
 				case NB_HEURES:
 				case PERIODE_HEURES:
@@ -102,7 +102,7 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 					break;
 
 			}
-			
+
 			// If quantity is 0, remove this record from the list of Sppprm
 			if (prm.getNbPrime() == 0) {
 				modifiedOrAddedSppprm.remove(prm);
@@ -110,39 +110,39 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 			}
 
 		}
-		
+
 		return modifiedOrAddedSppprm;
-		
+
 	}
-	
-	protected Sppprm findOrCreateSppprmRecord(List<Sppprm> existingRecords, Integer idAgent, Date dateJour, Integer norubr, Integer qte) {
-		
+
+	protected Sppprm findOrCreateSppprmRecord(List<Sppprm> existingRecords, Integer idAgent, Date dateJour,
+			Integer norubr, Integer qte) {
+
 		Sppprm prm = null;
-		
+
 		Integer nomatr = helperService.getMairieMatrFromIdAgent(idAgent);
 		Integer dateJourMairie = helperService.getIntegerDateMairieFromDate(dateJour);
-		
+
 		// First search through existing spacti
 		for (Sppprm a : existingRecords) {
-			if (a.getId().getNomatr().equals(nomatr) 
-				&& a.getId().getDatJour().equals(dateJourMairie)
-				&& a.getId().getNoRubr().equals(norubr)) {
+			if (a.getId().getNomatr().equals(nomatr) && a.getId().getDatJour().equals(dateJourMairie)
+					&& a.getId().getNoRubr().equals(norubr)) {
 				return a;
 			}
 		}
 
 		// Then Look for an exising record already existing in the DB
 		prm = exportPaieRepository.getSppprmForDayAgentAndNorubr(idAgent, dateJour, norubr);
-		
+
 		// At last create a new record, only if the qte is different than 0
 		if (prm == null && !qte.equals(0)) {
 			prm = new Sppprm();
 			prm.setId(new SppprmId(nomatr, dateJourMairie, norubr));
 		}
-		
+
 		if (prm != null)
 			existingRecords.add(prm);
-		
+
 		return prm;
 	}
 
@@ -150,48 +150,49 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 	public List<Spprim> exportPrimesMoisToPaie(List<VentilPrime> ventilPrimeOrderedByDateAsc) {
 
 		List<Spprim> pris = new ArrayList<Spprim>();
-		
+
 		for (VentilPrime ventilPrime : ventilPrimeOrderedByDateAsc) {
-			
+
 			if (ventilPrime.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPRIM)
 				continue;
-			
+
 			// Fetch or create Spprim
 			Spprim pri = findOrCreateSpprimmRecord(ventilPrime.getIdAgent(), ventilPrime.getDateDebutMois(),
 					ventilPrime.getRefPrime().getNoRubr(), ventilPrime.getQuantite());
-			
+
 			if (pri == null)
 				continue;
-			
+
 			// Fill in the number of Primes for the month
 			double qte = 0;
 			if (ventilPrime.getRefPrime().getTypeSaisie() == TypeSaisieEnum.NB_HEURES)
 				qte = helperService.convertMinutesToMairieNbHeuresFormat(ventilPrime.getQuantite());
 			else
 				qte = ventilPrime.getQuantite();
-			
-			pri.setMontantPrime(Math.ceil(qte));
-			
+
+			//#15317 on arrondi à l'unité supérieure
+			pri.setMontantPrime((int) Math.ceil(qte));
+
 			// Add the item to the list of hre modified/created
 			if (pri.getMontantPrime() != 0)
 				pris.add(pri);
 			else
 				exportPaieRepository.removeEntity(pri);
 		}
-		
+
 		return pris;
 	}
-	
+
 	protected Spprim findOrCreateSpprimmRecord(Integer idAgent, Date dateDebutMois, Integer norubr, Integer qte) {
-		
+
 		Spprim pri = null;
-		
+
 		Integer nomatr = helperService.getMairieMatrFromIdAgent(idAgent);
 		Integer dateDebMoisMairie = helperService.getIntegerDateMairieFromDate(dateDebutMois);
 
 		// Look for an exising record already existing in the DB
 		pri = exportPaieRepository.getSpprimForDayAgentAndNorubr(idAgent, dateDebutMois, norubr);
-		
+
 		// At last create a new record
 		if (pri == null && !qte.equals(0)) {
 			pri = new Spprim();
@@ -199,21 +200,22 @@ public class ExportPaiePrimeService implements IExportPaiePrimeService {
 			Date dateFinMois = new LocalDate(dateDebutMois).plusMonths(1).withDayOfMonth(1).toDate();
 			pri.setDateFin(helperService.getIntegerDateMairieFromDate(dateFinMois));
 		}
-		
+
 		return pri;
 	}
-	
+
 	@Override
 	public void deleteSppprmFromPrimesRejetees(List<Pointage> listPointageRejetesVentilesOrderedByDateAsc) {
-		
+
 		for (Pointage ptg : listPointageRejetesVentilesOrderedByDateAsc) {
-			
+
 			if (ptg.getTypePointageEnum() != RefTypePointageEnum.PRIME
 					|| ptg.getRefPrime().getMairiePrimeTableEnum() != MairiePrimeTableEnum.SPPPRM)
 				continue;
-			
+
 			// Then delete for an exising record already existing in the DB
-			exportPaieRepository.deleteSppprmForDayAndNorubr(ptg.getIdAgent(), ptg.getDateDebut(), ptg.getRefPrime().getNoRubr());
+			exportPaieRepository.deleteSppprmForDayAndNorubr(ptg.getIdAgent(), ptg.getDateDebut(), ptg.getRefPrime()
+					.getNoRubr());
 		}
 	}
 }
